pyproject.toml
[tool.poetry]
name = "hyperflask"
version = "0.1.3"
description = "Full stack web framework"
authors = ["Maxime Bouroumeau-Fuseau <maxime.bouroumeau@gmail.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/hyperflask/hyperflask"
packages = [{include = "hyperflask"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0"
flask-apispec-hyper = "^0.12"
flask-assets-pipeline = "^0.1"
flask-babel-hyper = "^5.0"
flask-collections = "^0.1"
flask-configurator = "^0.1"
flask-debugtoolbar = "^0.15"
flask-dramatiq = "^0.6"
flask-file-routes = "^0.1"
flask-files = "^0.1"
flask-geo = "^0.1"
flask-mailman-templates = "^0.1"
flask-mercure-sse = "^0.1"
#flask-otel = "^0.1"
flask-sqlorm = "^0.1"
flask-super-macros = "^0.1"
flask-talisman = "^1.1"
flask-wtf = "^1.2"
frozen-flask = "^1.0"
#dramatiq-sqlite = "^0.1"
gunicorn = "^22.0"
honcho = "^1.1"
htmx-flask = "^0.1"
jinja2-fragments = "^1.3"
jinja-layout = "^0.4"
jinja-wtforms = "^0.1"
periodiq = "^0.12"
python-slugify = "^8.0.4"
redis = "^5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
ruff = "^0.4.3"

[tool.poetry.scripts]
hyperflask = "hyperflask.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
